Added Question 8. Note freeing memory appears to be broken in Visual Studios debugger.

+++ Project Euler/Question_7.cpp
@@ -4,7 +4,7 @@
 
 int Question_7 () {
 	int Prime_Index = 0;
-	int bound = (10001 * (log (10001) + log (log (10001))));
+	int bound = (int)(10001 * (log (10001) + log (log (10001))));
 	int* Sieve_List = (int *) calloc ( bound, sizeof (int));
 	int i = 2;
 	while (Prime_Index < 10001) {
